Split code for JVM and Scala.js
