+++ project/plugin.sbt
@@ -1,8 +1,8 @@
-addSbtPlugin("com.github.gseitz" % "sbt-release" % "1.0.0")
+addSbtPlugin("com.github.gseitz" % "sbt-release" % "1.0.4")
 addSbtPlugin("org.xerial.sbt" % "sbt-sonatype" % "1.1")
 addSbtPlugin("com.jsuereth" % "sbt-pgp" % "1.0.0")
-//addSbtPlugin("io.get-coursier" % "sbt-coursier" % "1.0.0-M14")
-// sbt-scoverage doesn't work for Scala 2.12.0
-//addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.3.5")
-//addSbtPlugin("org.scoverage" % "sbt-coveralls" % "1.1.0")
+addSbtPlugin("io.get-coursier" % "sbt-coursier" % "1.0.0-M15")
+addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.5.0")
+addSbtPlugin("org.scoverage" % "sbt-coveralls" % "1.1.0")
+addSbtPlugin("org.scala-js" % "sbt-scalajs" % "0.6.14")
 scalacOptions ++= Seq("-deprecation", "-feature")
