--- src/test/resources/log-test.properties
@@ -1,1 +1,0 @@
-wvlet.log.test=DEBUG
