+++ wvlet-log/jvm/src/main/scala/wvlet/log/AsyncHandler.scala
