+++ wvlet-log/jvm/src/main/scala/wvlet/log/Guard.scala
