+++ wvlet-log/jvm/src/main/scala/wvlet/log/LogRotationHandler.scala
