+++ wvlet-log/jvm/src/main/scala/wvlet/log/io/Resource.scala
