+++ wvlet-log/jvm/src/test/resources/wvlet/log/invalid-loglevel.properties
