+++ wvlet-log/jvm/src/test/resources/wvlet/log/io/hello.txt
