+++ wvlet-log/shared/src/main/scala/wvlet/log/Handler.scala
