+++ wvlet-log/shared/src/main/scala/wvlet/log/LogFormat.scala
