+++ wvlet-log/shared/src/main/scala/wvlet/log/LogLevel.scala
