+++ wvlet-log/shared/src/main/scala/wvlet/log/LogMacros.scala
