+++ wvlet-log/shared/src/main/scala/wvlet/log/LogRecord.scala
