+++ wvlet-log/shared/src/main/scala/wvlet/log/LogSupport.scala
