+++ wvlet-log/shared/src/test/scala/wvlet/log/LoggerTest.scala
