+++ wvlet-log/shared/src/test/scala/wvlet/log/WvletLog.scala
